+++ svelte/sveltekit/src/components/Chat.svelte
